pyproject.toml
[project]
name = "metamed-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.3.0",
    "boto3>=1.26.0",
    "coverage>=7.0.0",
    "docling>=2.51.0",
    "fastapi>=0.100.0",
    "gunicorn>=21.0.0",
    "httpx>=0.24.0",
    "jwt>=1.4.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "langchain-pinecone>=0.2.12",
    "mailchimp-marketing>=3.0",
    "openai>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pinecone>=7.3.0",
    "postmark>=1.0",
    "psycopg2-binary>=2.9.6",
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.0.0",
    "pypdf2>=3.0.1",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-json-logger>=2.0.7",
    "python-memcached>=1.62",
    "python-multipart>=0.0.6",
    "redis>=5.0.0",
    "requests>=2.28.0",
    "slowapi>=0.1.8",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.23.0",
    "werkzeug>=3.1.5",
]
